Rename old empty state cell
